:boom: the scaffold now handles authentication
Instead of letting the application define the left "drawer", this release lets
the library define the drawer and applications can provide a list of `ListTile`
objects to customize it. The library creates the drawer header which now
displays the user's login state along with a button to login or logout,
depending on the state.