:bug: fix race condition
It seemed that we could never revoke the credentials. This code
revoked them in a microtask. However, futures don't make progress
until the main thread yeilds. Instead of yielding, we immediately
call `setState()` which `null`s the credentials before we revoke
them.

This version saves a reference so the microtask can actually
revoke the credentials.