:recycle: change name of query
The one-shot query is exactly the same as the streaming query except for
the `query` <-> `subscription` keyword. So I changed the keyword but
didn't rename the actual query. This resulted in a bunch of types with
equivalent names which required I rename the modules uniquely.

This commit fixes that mess.