:art: generalize middle portion of drawer
We had the middle portion as a `ListView`. This was a little too specific and 
constraining. Now it's just a widget that the developer provides and can be, essentially, 
anything. This widget is placed in a scrolling view so that the header and footer are 
always visible and a large widget will have to be scrolled to see all parts of it.